Waves!
There is now a very rough ability to construct waves of birds! Requires
editing a matrix of numbers in the Waves script, but it’s pretty
straightforward.

For this to be meaningful, it still needs bird AI. I held off on making
any real bird AI until we consulted together more about it.

Waves details:
The way it works is like this
pigeon = column 0
duck1 = column 1
duck2 = column 2
…
wave 1 = row 1
wave 2 = row 2
wave 3 = row 3

so that
3   0   0
0   3   0
1   2   3
means wave1 has 3 pigeons, wave2 has 3 duck1s, and wave 3 has 1 pigeon
+ 2duck1s + 3duck2s.

The trick with the system is that the birds are made in that order (all
pigeons first, then all duck1s…).

I would prefer a system where you can tell it the precise order of bird
after bird which one will appear next in the spawning sequence in any
given wave. So I could say “first, a duck1, then a pigeon, another
duck1, and finish with a duck2”