First Use of NameSpace!
Trying something new with coding out. Namespaces. I think it’s working.
I used it for the converting vectors to angles functions that I noticed
I was re-writing in many of the scripts. So now I can just reference it
from it’s own single spot.

Also started on the ducks flying V formation business. Seems quite
involved. And it’s late. so pause on that for now.