Bats!
I am pretty proud of these bats. If I thought I was proud of the
Pelicans (which, btw I’m not) then I’m crazy, but also these bats are
the shit.

they surround you and move around erratically, shifting between
randomized elliptical patterns. If you move too fast, one of them will
most certainly pop your B.

if you move slowly, they’ll follow you safely. just a matter of getting
rid of those fuckers. and it’s tough because of how unpredictable they
are.

Also, in making these I discovered the cause of BALLOON POPPING ERRORS!
It was happening because there was a delay on when to delete the
balloons that had popped, because i was keeping them around for the
duration of the popping noise and animation. So I set the balloon
script to nullify it’s reference to the balloon items, and there are NO
MORE TRANSFORM NOT FOUND ERRORS! Forever never more.

wait this one’s for bats… i’ll close with this then.

Bat mitzfahs frighten me. it’s time my enemies share my existential
dread.

haha so stupid, fuck it’s late