Rescaling Overhaul
Refactoring by scale. Unfortunately low fruit process for a lot of time
investment. It’s 11:11 am and I’m just going to sleep.

The big outcome is now everything can be scaled at a factor of 1
instead of 3.125 and 6.25. But for keeping the camera zoom at 1 (and
not 2 or 3) everything is scaled at x4 anyway. So feels kinda
worthless. But at least the background fits tight? It’s whatever…

Bats are animating now. Bird poo is fresh.

Still need to
1. Animate bird poo
2. Get multiple bird poos happening (unique render textures methinks)
3. Fix Crow sprites
4. Make a survival mode….
5. Create a kill tracking system

Hold off on fine tuning animals’ speeds and distances. Do this with
Nate and Hec.