WAVE UPGRADES
The Wave Class is fresh to death.

Now featuring powerful methods for mass bird production (think
delegates as inputs…)
Also starring the BirdWaiter class, which makes stop+go functionality a
breeze. Want to wait for all the birds to die before moving on? Easy.
Want to wait for 3 ducks to spawn before spawning a pelican? Equally
easy. Want to wait for 2 ducks to die before spawning 2 more AND wait
for 1 seagull remaining to spawn another seagull IN PARALLEL so either
can come first? It’s a snap, my man.

The point is, writing waves takes very little writing. The entire
pigeon wave script is like <40 lines, and there are at least 36 pigeons
produced in particular patterns with particular waiting schemes.

Points have been upgraded too! Now everything is worth 10x original
values (so pigeons are +10, not just +1).

Now to create
- a semi-random, endless wave generator
- a new bird! Woodpecker?
- bat/pecker/eagle/crow wave(s)
- bonus waves (like in galaga)

And to fix
- movement responsiveness
- throwing responsiveness
- points shown off screen

And to add
- mini tutorial (for movement, throwing, stabbing tentacles)