Movement Upgrades!
Now controls are **much** more RESPONSIVE.

Switched from Force-based movement to velocity-based movement, and the
control is much tighter. Still might need a way to come to a stop a bit
easier though. Maybe a tap near the center of the joystick?

Throws are instantaneous. The moment you let go is the moment the spear
flies, and it is MUCH more satisfying. Also, the time between throws
was cut in half. So you can throw twice as many spears in the same time!

Tentacles is working again too!

Still need to get the pelicans to redirect once they’ve been hit. Maybe
not though, now that you can throw with such speed.