Interchangable Weapons! + Tutorial
Now with a system for using different weapons
(Spear, Lightning, Mace)...

Lightning demo is fully functional too!

The idea here is to unlock these weapons as you complete the story mode

And indeed, there is a short tutorial. More of just some on-screen text,
but it works quite well.

New SFX for balloon popping, lightning strikes, and spear throws